editing jwt
