+++ server/package.json
@@ -33,6 +33,7 @@
     "jsonwebtoken": "^8.5.1",
     "mongoose": "^6.0.14",
     "morgan": "^1.10.0",
+    "nanoid": "^3.3.0",
     "nodemailer": "^6.7.2",
     "nodemailer-smtp-transport": "^2.7.4",
     "nodemon": "^2.0.15",
