+++ server/src/routes/auth/linkedinRouter.ts
@@ -1,6 +1,6 @@
 import { Router } from "express";
 import passport from "passport";
-import { signIn, logout } from "../../controllers/auth/passport";
+import Auth from "../../controllers/auth";
 
 const router = Router();
 
@@ -9,9 +9,7 @@ router.get("/", passport.authenticate("linkedin", { state: "SOME STATE" }));
 router.get(
   "/callback",
   passport.authenticate("linkedin", { failureRedirect: "auth/failed" }),
-  signIn
+  Auth.login
 );
-
-router.get("/logout", logout);
 
 export default router;
